Change strategy to launch mobilization for use reusable components. Issue #497
