+++ routes/admin/authenticated/sidebar/index.js
@@ -23,7 +23,6 @@ export default store => ({
         require('./community-settings-recipient').default(store),
         require('./community-settings-report').default(store),
         require('./mobilizations-edit').default(store),
-        require('./mobilizations-launch').default(store),
         require('./mobilizations-list').default(store),
         require('./mobilizations-new').default(store),
         require('./mobilizations-settings-analytics').default(store),
