--- routes/admin/authenticated/sidebar/mobilizations-launch/index.js
@@ -1,11 +1,0 @@
-// polyfill webpack require.ensure
-if (typeof require.ensure !== 'function') require.ensure = (d, c) => c(require)
-
-export default store => ({
-  path: 'mobilizations/:mobilization_id/launch',
-  getComponent (nextState, callback) {
-    require.ensure([], function (require) {
-      callback(null, require('./page.connected').default)
-    })
-  }
-})
