--- routes/admin/authenticated/sidebar/mobilizations-launch/page.connected.js
@@ -1,39 +1,0 @@
-import { provideHooks } from 'redial'
-import { connect } from 'react-redux'
-import { reduxForm } from 'redux-form'
-
-import MobSelectors from '~client/mobrender/redux/selectors'
-import * as MobActions from '~client/mobrender/redux/action-creators'
-import { fields, validate } from '~mobilizations/components/form-custom-domain'
-
-import Page from './page'
-
-const redial = {
-  fetch: ({ dispatch, getState, params }) => {
-    const state = getState()
-    const promises = []
-
-    !MobSelectors(state).hasCurrentMobilization() && promises.push(
-      dispatch(MobActions.selectMobilization(params.mobilization_id))
-    )
-    return Promise.all(promises)
-  }
-}
-
-const mapStateToProps = state => {
-  const mobilization = MobSelectors(state).getMobilization()
-  return {
-    initialValues: mobilization,
-    mobilization
-  }
-}
-
-const mapActionCreatorsToProps = {
-  submit: MobActions.asyncUpdateMobilization
-}
-
-export default provideHooks(redial)(
-  connect(mapStateToProps, mapActionCreatorsToProps)(
-    reduxForm({ form: 'mobilizationLaunchDomainForm', fields, validate })(Page)
-  )
-)
